online_inference/requirements.txt
attrs==21.2.0
cycler==0.10.0
iniconfig==1.1.1
joblib==1.0.1
kiwisolver==1.3.1
marshmallow==3.11.1
marshmallow-dataclass==8.4.1
matplotlib
mypy-extensions==0.4.3
numpy
packaging==20.9
pandas
Pillow==8.2.0
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
scikit-learn==0.24.2
scipy
seaborn==0.11.1
six==1.16.0
threadpoolctl==2.1.0
toml==0.10.2
typing-extensions==3.10.0.0
typing-inspect==0.6.0
fastapi
uvicorn
requests
docker
